Panic commit
I found LASM performs very poor in barotropic experiment. The tracer should not
be too filament, because in this case the linear approximation will be violated
(very noise in the geopotential depth field). So I tried to keep the tracers
compact, and the results are better, but this algorithm has not been polished.

Next I will modify the tracer split and merge algorithm by directly send mass
from the tracer that is judged to be too long to its surrounding tracers
without introducing intermediate tracers, and the amount of sending mass should
be small at one time and proportional to the filament degree of the tracer.