Revised interpacel mixing algorithm
I have replaced the interparcel mixing algorithm. In the new algorithm
the parcel is judged to be mixed or not by the skeleton point position
bias directly. The extreme large and small parcels in the original one
are avoided. The mixing manner is also changed. The most important
difference is that the neighbor parcel shapes are not changed, and the
problematic parcel volume is also not changed. The density of the
neighbor tracers is restored to a weighted mean density, and the density
of the problematic tracer is left to be set according to mass
conservation requirement.

I also add smoothing to the geopotential depth tracer in barotropic test
case.